Correct an actual bug.  The warning was: "comparison is always false due to limited range of data type".  The variable remBytes was declared as a uint16_t but then compared to a const of 300000 which is bigger then can be stored in a uint16_t, the correction is to make remBytes a uint32_t which is big enough to actually have 300000 stored in it. It is quite clear that the number 300000 was intentional as it is used at least 6 places in code.
Really I would recomend replacing the 6 "300000" with a const
uint32_t offsetSize = 300000 and then use the const variable in the 6
locations much less error prone in the long term.

+++ Code/Adafruit_FONA.cpp
@@ -2391,7 +2391,7 @@ boolean Adafruit_FONA::FTP_PUT(const char* fileName, const char* filePath, char*
   // Use extended PUT method if there's more than 1024 bytes to send
   if (numBytes >= 1024) {
     // Repeatedly PUT data until all data is sent
-    uint16_t remBytes = numBytes;
+    uint32_t remBytes = numBytes;
     uint16_t offset = 0; // Data offset
     char sendArray[strlen(content)+1];
     strcpy(sendArray, content);
